requirements.txt
auto_mix_prep==0.2.0
click==8.1.3
hydra-core==1.3.1
matplotlib==3.6.3
numpy==1.24.1
pytest==7.2.1
python-dotenv==0.21.0
setuptools==65.6.3
torch==1.13.1
torch_geometric==2.2.0
wandb==0.13.9

# local package
-e .